pyproject.toml
[project]
name = "portableagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.31.0",
    "celery>=5.6.2",
    "fastapi>=0.128.7",
    "fastmcp>=2.14.5",
    "gcsa>=2.6.0",
    "google-api-python-client>=2.189.0",
    "google-auth>=2.48.0",
    "google-auth-oauthlib>=1.2.4",
    "loguru>=0.7.3",
    "mcp>=1.26.0",
    "python-dotenv>=1.2.1",
    "sqlalchemy[asyncio]>=2.0.46",
]
